SAL foreach,  MsgQ search, Approval part of Proto
The sorted Address list now can have the foreach construction used on
it. We can search the message queues for messages with particular
content. The byzantium protocol is in early works. A helper function to
convert a string containing addresses to a sorted address list has been
made. Bug hunts are necessary.